+++ backend/src/main/java/org/neuefische/applicationmangementapp/service/ApplicationService.java
@@ -1,6 +1,6 @@
 package org.neuefische.applicationmangementapp.service;
 
-import org.neuefische.applicationmangementapp.execaptions.NoSuchId;
+import org.neuefische.applicationmangementapp.exceptions.NoSuchId;
 import org.neuefische.applicationmangementapp.model.Application;
 
 import org.neuefische.applicationmangementapp.model.ApplicationDtoForCreated;
